Silently ignore overrides of Maxis ordinances
The 140-ordinances folder we use may contain mods that override the
Maxis ordinances.
The Maxis ordinances are handled by the game's built-in GZCOM classes.

As a performance optimization, we skip reading the exemplar data when
the group and instance ids match a Maxis ordinance exemplar.
This saves a bunch of unnecessary work on our DLL's startup code path.

We report an error in the log file if a custom ordinance has the same
instance id as a Maxis ordinance with a different group id, but this
should almost never occur.
GZCOM requires the class ids to be unique, and the ordinance system
uses the instance id as the class id.